requirements.txt
tweepy>=4.14.0
openai>=1.14.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
pillow>=10.3.0
vaderSentiment>=3.3.2
pytz>=2024.1
